+++ example/lib/payment_example.dart
@@ -4,18 +4,22 @@ import 'package:flutter/material.dart';
 class PaymentExample extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
+    final controller = PaymentController();
     return PaymentForm(
+      controller: controller,
       decorationElement: OutlineDecorationElement(),
-      actionPayment: (cardNumber, cvv, date) async {
-        print("Credit Card information : $cardNumber,$cvv,$date");
-      },
       errorMessageCVV: "cvv is invalid",
       errorMessageDateExpiration: "date expiration is invalid",
       errorIsRequiredMessage: "This field  is required",
       labelCVV: "cvv",
       labelDateExpiration: "date expiration",
       labelCardNumber: "card number",
-      paymentText: Text("purchase"),
+      submitButton: ElevatedButton(
+        onPressed: () {
+          controller.validate();
+        },
+        child: Text("pay"),
+      ),
     );
   }
 }