+++ lib/src/widgets/cvv_text_field.dart
@@ -6,27 +6,61 @@ import '../utilities/constants.dart';
 class CvvTextField extends StatelessWidget {
   final TextEditingController controller;
   final CVVElement element;
-  final FocusNode currentFocus;
-  final FocusNode nextFocus;
+  final FocusNode? currentFocus;
+  final FocusNode? nextFocus;
+  final ValueNotifier<String?>? errorNotifier;
 
   CvvTextField({
-    this.controller,
-    this.element,
+    required this.controller,
+    required this.element,
     this.currentFocus,
     this.nextFocus,
-    Key key,
+    this.errorNotifier,
+    Key? key,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    if (element.initValue.isNotEmpty) controller.text = element.initValue;
+    if (element.initValue != null && element.initValue!.isNotEmpty)
+      controller.text = element.initValue!;
+    if (errorNotifier != null) {
+      return ValueListenableBuilder<String?>(
+        valueListenable: errorNotifier!,
+        builder: (ctx, error, _) {
+          return TextFormField(
+            controller: controller,
+            validator: element.validator,
+            keyboardType: Constants.getInput(element.typeInput),
+            readOnly: element.readOnly,
+            enabled: true,
+            onTap: element.onTap as void Function()?,
+            focusNode: currentFocus,
+            obscureText: true,
+            textInputAction:
+                nextFocus == null ? TextInputAction.done : TextInputAction.next,
+            onFieldSubmitted: (v) {
+              Constants.fieldFocusChange(context, currentFocus, nextFocus);
+            },
+            decoration:
+                Constants.setInputBorder(context, element.decorationElement)
+                    .copyWith(
+              labelText: element.label,
+              hintText: element.hint,
+              errorText: error,
+              enabled: true,
+              suffixIcon: null,
+            ),
+          );
+        },
+      );
+    }
     return TextFormField(
       controller: controller,
       validator: element.validator,
       keyboardType: Constants.getInput(element.typeInput),
       readOnly: element.readOnly,
       enabled: true,
-      onTap: element.onTap,
+      onTap: element.onTap as void Function()?,
       focusNode: currentFocus,
       obscureText: true,
       textInputAction: